+++ src/pages/Edital/EnrolledList/Service.tsx
@@ -1,0 +1,12 @@
+import api from "../../../services/Api";
+import { IADetalhes, IInscritos } from "./Interfaces";
+
+const getEnrolledList = (editalId: string) => {
+  return api.get<IADetalhes[]>(`/processos-seletivos/${editalId}/inscricoes`);
+};
+
+const getEnrolledDetails = (editalId: string | undefined) => {
+  return api.get<IInscritos>(`/processos-seletivos/${editalId}`);
+};
+
+export { getEnrolledDetails, getEnrolledList };
